Move pkg/kubectl/describe to staging
