+++ pkg/kubectl/BUILD
@@ -84,7 +84,6 @@ filegroup(
         ":package-srcs",
         "//pkg/kubectl/apply:all-srcs",
         "//pkg/kubectl/cmd:all-srcs",
-        "//pkg/kubectl/describe:all-srcs",
         "//pkg/kubectl/explain:all-srcs",
         "//pkg/kubectl/generate:all-srcs",
         "//pkg/kubectl/metricsutil:all-srcs",
