+++ pkg/kubectl/cmd/describe/describe.go
@@ -28,11 +28,11 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/cli-runtime/pkg/resource"
+	"k8s.io/kubectl/pkg/describe"
+	describeversioned "k8s.io/kubectl/pkg/describe/versioned"
 	"k8s.io/kubectl/pkg/util/i18n"
 	"k8s.io/kubectl/pkg/util/templates"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
-	"k8s.io/kubernetes/pkg/kubectl/describe"
-	describeversioned "k8s.io/kubernetes/pkg/kubectl/describe/versioned"
 )
 
 var (
