+++ pkg/kubectl/cmd/describe/describe_test.go
@@ -26,10 +26,10 @@ import (
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/cli-runtime/pkg/resource"
 	"k8s.io/client-go/rest/fake"
+	"k8s.io/kubectl/pkg/describe"
+	versioneddescribe "k8s.io/kubectl/pkg/describe/versioned"
 	"k8s.io/kubectl/pkg/scheme"
 	cmdtesting "k8s.io/kubernetes/pkg/kubectl/cmd/testing"
-	"k8s.io/kubernetes/pkg/kubectl/describe"
-	versioneddescribe "k8s.io/kubernetes/pkg/kubectl/describe/versioned"
 )
 
 // Verifies that schemas that are not in the master tree of Kubernetes can be retrieved via Get.
