+++ staging/src/k8s.io/kubectl/BUILD
@@ -10,6 +10,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//staging/src/k8s.io/kubectl/pkg/apps:all-srcs",
+        "//staging/src/k8s.io/kubectl/pkg/describe:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/drain:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/generated:all-srcs",
         "//staging/src/k8s.io/kubectl/pkg/rawhttp:all-srcs",
