+++ staging/src/k8s.io/kubectl/pkg/describe/BUILD
@@ -3,7 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["interface.go"],
-    importpath = "k8s.io/kubernetes/pkg/kubectl/describe",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/kubectl/pkg/describe",
+    importpath = "k8s.io/kubectl/pkg/describe",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/apimachinery/pkg/api/meta:go_default_library",
@@ -22,7 +23,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//pkg/kubectl/describe/versioned:all-srcs",
+        "//staging/src/k8s.io/kubectl/pkg/describe/versioned:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
